requirements.txt
pandas
openpyxl # for Excel
python-dotenv # for .env
requests # for API
pyyaml # for YAML
inquirer # for CLI
fastapi  # for API
uvicorn # for API